Vida
Health Manager:
Incluye Escudo y Vida
Se puede usar para IA y Jugadores

Cambios al Disparo:
Ahora es como metralleta

+++ 3dshooter/Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,4 +1,4 @@
 public interface IDamageable
 {
-    void ApplyDamage(int damageTaken);
+    void ApplyDamage(float damageTaken);
 }
